analysis page, expenses analysis, top spending category, category breakdown for expenses by daily, weekly, etc

+++ lib/features/wallet/presentation/providers/wallet_notifier.dart
@@ -3,7 +3,7 @@ import 'package:finance_management/features/auth/presentation/providers/auth_pro
 import 'package:finance_management/features/wallet/application/wallet_service.dart';
 import 'package:finance_management/features/wallet/domain/wallet.dart';
 import 'package:finance_management/features/wallet/presentation/providers/wallet_state.dart';
-import 'package:flutter/material.dart'; // Tambahkan untuk akses Icons default
+import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_riverpod/legacy.dart';
 
@@ -13,7 +13,6 @@ class WalletNotifier extends StateNotifier<WalletState> {
 
   WalletNotifier(this._service, this._ref) : super(WalletState());
 
-  // Method Save Tunggal (Bisa Add maupun Update)
   Future<void> saveWallet({
     String? id,
     required String name,
@@ -22,13 +21,14 @@ class WalletNotifier extends StateNotifier<WalletState> {
   }) async {
     try {
       state = state.copyWith(isLoading: true);
-      final userId = _ref.read(authNotifierProvider).user!.id;
+
+      final userId = _ref.read(authStateChangesProvider).value?.uid;
+      if (userId == null) throw Exception("User not authenticated");
 
       final wallet = Wallet(
         id: id ?? '',
         name: name,
         balance: balance,
-        // Gunakan iconCode yang dikirim atau default ke wallet icon
         iconCode: iconCode ?? Icons.account_balance_wallet.codePoint,
       );
 
@@ -44,11 +44,13 @@ class WalletNotifier extends StateNotifier<WalletState> {
     }
   }
 
-  // Tetap sediakan update khusus jika hanya ingin passing objek Wallet langsung
   Future<void> updateWallet(Wallet wallet) async {
     try {
       state = state.copyWith(isLoading: true);
-      final userId = _ref.read(authNotifierProvider).user!.id;
+
+      final userId = _ref.read(authStateChangesProvider).value?.uid;
+      if (userId == null) throw Exception("User not authenticated");
+
       await _service.updateWallet(userId, wallet);
       state = state.copyWith(isLoading: false, errorMessage: null);
     } catch (e) {
@@ -59,7 +61,10 @@ class WalletNotifier extends StateNotifier<WalletState> {
   Future<void> deleteWallet(String walletId) async {
     try {
       state = state.copyWith(isLoading: true);
-      final userId = _ref.read(authNotifierProvider).user!.id;
+
+      final userId = _ref.read(authStateChangesProvider).value?.uid;
+      if (userId == null) throw Exception("User not authenticated");
+
       await _service.deleteWallet(userId, walletId);
       state = state.copyWith(isLoading: false, errorMessage: null);
     } catch (e) {
